Handle auto-versioning
A request for `home/text.md` can return a URI ending with for example
`home/text.md#1` if the user has edited the node.

This commit adds support for using auto-versioning. If a response node
is returned with a version, it is also used for unversioned URIs.

This also makes `djedi.fetchMany` non-public (`djedi._fetchMany`)
because the versioning stuff made me realize that it doesn’t make sense.
The user would be able to fetch several nodes at once, but not know
which part of the response corresponds to which input node.