Improve reloading of the Djedi sidebar
- Remove old iframe before inserting a new one. This is nice when hot
reloading code while the sidebar is open.

- Always reload the iframe when a node is unmounted, just like we always
reload when a node is mounted. Previously, when a node was unmounted the
iframe was only reloaded when the node in question was the last rendered
instance of its URI.

- Use a `MutationObserver` to detect when nodes are rendered into the
page. This is needed when a `<Node>` uses the `render` prop to render
`null` initially. After some time it might choose to render then
`<span>`. But the node is already reported as rendered. By using a
`MutationObserver` to detect when `<span data-i18n>` are added/removed
we can make sure that their outlines are added/removed as well.