Update md editor preview on the fly
Previously, it was only updated when leaving the text area. Now it is
updated when you stop typing for 400 ms. I thought for a long time that
there the md editor did not support updating the preview, because it
never occurred to me that I could unfocus the text area to preview. This
should make the feature more obvious.