Use only one scrollbar in the sidebar nodes listing
Previously, there was one scrollbar for every section. That was very
hard spot, especially since the max-height of every section exactly
matched a whole number of nodes. Also, it was pretty awkward scrolling
in those little boxes. Finally, if there were more sections than could
fit on your screen, the ones off-screen couldn't be reached. This commit
switches to a single scrollbar for the entire thing, which feels more
natural.