Silence annoying 403 errors in browser dev tools
This happens when trying to inject the admin via djedi-react. The API
responds with 403 if not logged in as an admin. While that might be an
appropriate status code, it is annoying to get red errors in the browser
console because of that “failed request.” Especially if you’re not super
familiar with djedi internals it looks like something is wrong. This
commit responds with 204 No Content instead.